solution/scraper/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
black==24.8.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pdfminer.six==20191110
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pyodbc==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
sniffio==1.3.1
sqlalchemy==2.0.32
starlette==0.38.4
textract==1.6.5
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6